ios version(16 -> 15) & add 페이지 main페이지 +버튼으로 이동
