+++ NyamNyam/Model/Wish.swift
@@ -27,3 +27,4 @@ class Wish{
     }
     
 }
+
